Delete RequestExt.kt and move logic into CurlLoggingStrategy class
