+++ src/test/kotlin/io/github/rybalkinsd/kohttp/interceptors/logging/CurlLoggingStrategyTest.kt
@@ -1,4 +1,4 @@
-package io.github.rybalkinsd.kohttp.ext
+package io.github.rybalkinsd.kohttp.interceptors.logging
 
 import okhttp3.FormBody
 import okhttp3.Request
@@ -8,12 +8,12 @@ import kotlin.test.assertEquals
 /**
  * @author doyaaaaaken
  */
-class RequestExtTest {
+class CurlLoggingStrategyTest {
 
     @Test
     fun `build curl command of simple get request`() {
         val request = Request.Builder().url("https://postman-echo.com/get").build()
-        val actual = request.buildCurlCommand()
+        val actual = CurlLoggingStrategy().buildCurlCommand(request)
 
         assertEquals("curl -X GET \"https://postman-echo.com/get\"", actual)
     }
@@ -25,7 +25,7 @@ class RequestExtTest {
                 .header("Cookie", "foo=6; bar=28")
                 .header("Content-Type", "\"application/json\"")
                 .build()
-        val actual = request.buildCurlCommand()
+        val actual = CurlLoggingStrategy().buildCurlCommand(request)
 
         assertEquals("curl -X GET -H \"Cookie: foo=6; bar=28\" -H \"Content-Type: application/json\" \"https://postman-echo.com/get\"", actual)
     }
@@ -36,7 +36,7 @@ class RequestExtTest {
                 .url("https://postman-echo.com/post")
                 .post(FormBody.Builder().add("foo", "123").add("bar", "abc").build())
                 .build()
-        val actual = request.buildCurlCommand()
+        val actual = CurlLoggingStrategy().buildCurlCommand(request)
 
         assertEquals("curl -X POST --data \$'foo=123&bar=abc' \"https://postman-echo.com/post\"", actual)
     }
